% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{adddefinition}
\alias{adddefinition}
\title{adddefinition}
\usage{
adddefinition(term, definition, linktext1 = NULL, linktext2 = NULL)
}
\arguments{
\item{term}{Term to be defined}

\item{definition}{Definition of term}

\item{linktext1}{Text to appear in place of link associated with definition (optional)}

\item{linktext2}{Link associated with definition (optional)}
}
\value{
A dataframe containing all information associated to definitions.
}
\description{
Add a definition of a term relevant to the workbook.
}
\details{
adddefinition function will add a definition and its description to the workbook, specifically in the definitions worksheet.
Add definitions if wanted, if not then do run the adddefinition function.
term and definition are compulsory parameters.
A link can be added with each definition.
linktext1 and linktext2: linktext1 should be the text you want to appear and linktext2 should be the underlying link to a website, file etc.
}
