% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{addnote}
\alias{addnote}
\title{addnote}
\usage{
addnote(
  notenumber,
  notetext,
  applictabtext = NULL,
  linktext1 = NULL,
  linktext2 = NULL
)
}
\arguments{
\item{notenumber}{Note number}

\item{notetext}{Note description}

\item{applictabtext}{Table(s) a note is applicable to (optional)}

\item{linktext1}{Text to appear in place of a link associated to a note (optional)}

\item{linktext2}{Link associated to a note (optional)}
}
\value{
A dataframe containing all information associated to notes
}
\description{
Add a note to a particular table or tables
}
\details{
addnote function will add a note and its description to the workbook, specifically in the notes worksheet.
Add notes if wanted, if not then do not run the addnote function.
A link can be provided with each note as well a list of tables that the note applies to.
notenumber and notetext are the only compulsory parameters.
All other parameters are optional and preset to NULL, so only need to be defined if they are wanted.
applictabtext should be set to a vector of sheet names if a column is wanted which lists which worksheets a note is applicable to.
linktext1 and linktext2: linktext1 should be the text you want to appear and linktext2 should be the underlying link to a website, file etc.
}
