% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{savingtables}
\alias{savingtables}
\title{savingtables}
\usage{
savingtables(filename, odsfile = "No", deletexlsx = NULL)
}
\arguments{
\item{filename}{File path and file name of final output, including file type}

\item{odsfile}{Define whether to convert output to an ods file (optional)}

\item{deletexlsx}{Define whether to delete the xlsx file output (optional)}
}
\value{
A workbook saved to the network drive
}
\description{
Saving the final output
}
\details{
The savingtables function only requires that the location and name of the spreadsheet be specified.
A xls file can be saved but it is recommended to use a xlsx file instead.
Cannot save directly to ods, instead first a xlsx file is saved and then converted.
Default setting is not to create the ods file from the xlsx file.
If only xlsx file is wanted keep odsfile = "No" but if both ods and xlsx file wanted set odsfile = "Yes" and deletexlsx = "No".
If only the ods file wanted set odsfile = "Yes" and deletexlsx = "Yes".
}
