% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{workbook}
\alias{workbook}
\title{workbook}
\usage{
workbook(
  covertab = NULL,
  contentstab = NULL,
  notestab = NULL,
  autonotes = NULL,
  definitionstab = NULL,
  fontnm = "Arial",
  fontcol = "black",
  fontsz = 12,
  fontszst = 14,
  fontszt = 16,
  title = NULL,
  creator = NULL,
  subject = NULL,
  category = NULL
)
}
\arguments{
\item{covertab}{Define whether a cover page is required  (optional)}

\item{contentstab}{Define whether a contents page is required (optional)}

\item{notestab}{Define whether a notes page is required (optional)}

\item{autonotes}{Define whether automated listing of notes associated with a table is required (optional)}

\item{definitionstab}{Define whether a definitions page is required (optional)}

\item{fontnm}{Define the font name used in the final output (optional)}

\item{fontcol}{Define the font colour used in the final output (optional)}

\item{fontsz}{Define the general font size used in the final output (optional)}

\item{fontszst}{Define the font size for subtitles used in the final output (optional)}

\item{fontszt}{Define the font size for titles used in the final output (optional)}

\item{title}{Define the title to go into the document information in the final output (optional)}

\item{creator}{Define the creator to go into the document information in the final output (optional)}

\item{subject}{Define the subject to go into the document information in the final output (optional)}

\item{category}{Define the category to go into the document information in the final output (optional)}
}
\value{
A workbook called wb will appear in the global environment. Necessary R packages will be installed.
}
\description{
Create the openxlsx workbook where data tables are added.
}
\details{
The workbook function creates a new workbook with the required metadata worksheets and defines the workbook's font name, colour and sizes.
All parameters are optional and preset.
If a cover page, contents page, notes page or definitions page are required then set the parameter to "Yes" when calling the function.
autonotes is required if a line is wanted towards the top of the worksheet which lists all the note numbers associated with the worksheet (set to "Yes" if wanted).
Default font is Arial with a black colour and size ranging from 12 to 16 - change if want to when calling the function.
title, creator, subject and category refer to the document information properties displayed in the final Excel workbook.
}
