% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibletablesR-package.R
\docType{package}
\name{accessibletablesR-package}
\alias{accessibletablesR}
\alias{accessibletablesR-package}
\title{accessibletablesR: Creating accessible tables in an Excel workbook}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates Excel spreadsheets which align as closely as possible to UK Government Statistical Service's accessibility criteria. Relies upon the R package 'openxlsx'. The package was developed using R 4.1.3, 'tidyverse' version 2.0.0, 'openxlsx' version 4.2.5.2, 'devtools' version 2.4.5, 'conflicted' version 1.2.0 and 'odsconvertr' version 0.2.2. Alternative options include the R package 'a11ytables' and the Python package 'gptables'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/minifiemj/accessibletablesR}
}

}
\author{
\strong{Maintainer}: Matthew Minifie \email{matthew.minifie@ons.gov.uk}

}
\keyword{internal}
