% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{contentstable}
\alias{contentstable}
\title{contentstable}
\usage{
contentstable(gridlines = "Yes", colwid_spec = NULL, extracols = NULL)
}
\arguments{
\item{gridlines}{Define whether gridlines are present (optional)}

\item{colwid_spec}{Define widths of columns (optional)}

\item{extracols}{Define whether additional columns required (optional)}
}
\value{
A worksheet with a contents page of tables in the workbook.
}
\description{
Create a contents page for the workbook.
}
\details{
contentstable function creates a table of contents for the workbook.
If no contents page wanted, then do not run the contentstable function.
gridlines is by default set to "Yes", change to "No" if gridlines are not wanted.
Column widths are automatically set unless user defines specific values in colwid_spec.
Extra columns can be added, need to set extracols to "Yes" and create a dataframe extracols_contents with the desired extra columns.
}
