% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{definitionstab}
\alias{definitionstab}
\title{definitionstab}
\usage{
definitionstab(
  contentslink = NULL,
  gridlines = "Yes",
  colwid_spec = NULL,
  extracols = NULL
)
}
\arguments{
\item{contentslink}{Define whether a link to the contents page is wanted (optional)}

\item{gridlines}{Define whether gridlines are present (optional)}

\item{colwid_spec}{Define widths of columns (optional)}

\item{extracols}{Define whether additional columns required (optional)}
}
\value{
A worksheet of the definitions page for the workbook.
}
\description{
Create a definitions page for the workbook.
}
\details{
definitionstab creates the worksheet with information on definitions.
If definitions not wanted, then do not run the definitionstab function.
There are three parameters and they are optional and preset. Change contentslink to "No" if you want a contents tab but do not want a link to it in the definitions tab. Change gridlines to "No" if gridlines are not wanted.
Column widths are automatically set but the user can specify the required widths in colwid_spec.
Extra columns can be added by setting extracols to "Yes" and creating a dataframe extracols_definitions with the desired extra columns.
}
