% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{notestab}
\alias{notestab}
\title{notestab}
\usage{
notestab(
  contentslink = NULL,
  gridlines = "Yes",
  colwid_spec = NULL,
  extracols = NULL
)
}
\arguments{
\item{contentslink}{Define whether a link to the contents page is wanted (optional)}

\item{gridlines}{Define whether gridlines are present (optional)}

\item{colwid_spec}{Define widths of columns (optional)}

\item{extracols}{Define whether additional columns required (optional)}
}
\value{
A worksheet of the notes page for the workbook.
}
\description{
Create a notes page for the workbook.
}
\details{
notestab function will create a notes worksheet in the workbook and includes notes added using the addnote function.
If notes not wanted, then do not run the notestab function.
There are three parameters and they are optional and preset. Change contentslink to "No" if you want a contents tab but do not want a link to it in the notes tab. Change gridlines to "No" if gridlines are not wanted.
Column widths are automatically set but the user can specify the required widths in colwid_spec.
Extra columns can be added by setting extracols to "Yes" and creating a dataframe extracols_notes with the desired extra columns.
}
