% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{notestab}
\alias{notestab}
\title{accessibletablesR::notestab}
\usage{
notestab(
  contentslink = NULL,
  gridlines = "Yes",
  colwid_spec = NULL,
  extracols = NULL
)
}
\arguments{
\item{contentslink}{Define whether a link to the contents page is wanted (optional)}

\item{gridlines}{Define whether gridlines are present (optional)}

\item{colwid_spec}{Define widths of columns (optional)}

\item{extracols}{Define whether additional columns required (optional)}
}
\value{
A worksheet of the notes page for the workbook.
}
\description{
Create a notes page for the workbook.
}
\details{
notestab function will create a notes worksheet in the workbook and includes notes added using the addnote function.
If notes not wanted, then do not run the notestab function.
There are three parameters and they are optional and preset. Change contentslink to "No" if you want a contents tab but do not want a link to it in the notes tab. Change gridlines to "No" if gridlines are not wanted.
Column widths are automatically set but the user can specify the required widths in colwid_spec.
Extra columns can be added by setting extracols to "Yes" and creating a dataframe extracols_notes with the desired extra columns.
}
\examples{
accessibletablesR::workbook(covertab = "Yes", contentstab = "Yes", notestab = "Yes", definitionstab = "Yes",
                            autonotes = "Yes", title = "Fuel consumption and aspects of car design and performance for various cars",
                            creator = "An organisation")
                            
accessibletablesR::creatingtables(title = "Fuel consumption and aspects of car design and performance for various cars C",
                                  subtitle = "Cars",
                                  extraline1 = "Link to contents",
                                  extraline2 = "Link to notes",
                                  extraline3 = "Link to definitions",
                                  sheetname = "Table_3", table_data = dummydf, tablename = "thirdtable", headrowsize = 40,
                                  numdatacols = c(2:8,11:13), numdatacolsdp = c(1,0,1,0,2,1,2,0,0,3),
                                  othdatacols = c(9,10), columnwidths = "specified",
                                  colwid_spec = c(18,18,18,15,17,15,12,17,12,13,23,22,12))
                                  
accessibletablesR::contentstable()

accessibletablesR::addnote(notenumber = "note1", notetext = "Google is an internet search engine", applictabtext = "All", linktext1 = "Google", linktext2 = "https://www.ons.google.co.uk") 

accessibletablesR::notestab()

accessibletablesR::adddefinition(term = "Usual resident", definition = "A usual resident is anyone who, on Census Day, 21 March 2021 was in the UK and had stayed or intended to stay in the UK for a period of 12 months or more, or had a permanent UK address and was outside the UK and intended to be outside the UK for less than 12 months.")

accessibletablesR::definitionstab()

accessibletablesR::coverpage(title = "Fuel consumption and aspects of car design and performance for various cars",
                             intro = "Some made up data about cars",
                             about = "The output of an example of how to use accessibletablesR",
                             source = "R mtcars",
                             relatedlink = "https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/mtcars)",
                             relatedtext = "mtcars: Motor trend car road tests",
                             dop = "26 October 2023",
                             blank = "There should be no blank cells",
                             names = "Your name",
                             email = "yourname@emailprovider.com",
                             phone = "01111 1111111111111",
                             reuse = "Yes", govdept = NULL)
                             
accessibletablesR::savingtables("D:/mtcarsexample.xlsx", odsfile = "Yes", deletexlsx = "No")

}
