% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible_tables.R
\name{savingtables}
\alias{savingtables}
\title{accessibletablesR::savingtables}
\usage{
savingtables(filename, odsfile = "No", deletexlsx = NULL)
}
\arguments{
\item{filename}{File path and file name of final output, including file type}

\item{odsfile}{Define whether to convert output to an ods file (optional)}

\item{deletexlsx}{Define whether to delete the xlsx file output (optional)}
}
\value{
A workbook saved to the network drive
}
\description{
Saving the final output
}
\details{
The savingtables function only requires that the location and name of the spreadsheet be specified.
A xls file can be saved but it is recommended to use a xlsx file instead.
Cannot save directly to ods, instead first a xlsx file is saved and then converted.
Default setting is not to create the ods file from the xlsx file.
If only xlsx file is wanted keep odsfile = "No" but if both ods and xlsx file wanted set odsfile = "Yes" and deletexlsx = "No".
If only the ods file wanted set odsfile = "Yes" and deletexlsx = "Yes".
}
\examples{
accessibletablesR::workbook(covertab = "Yes", contentstab = "Yes", notestab = "Yes", definitionstab = "Yes",
                            autonotes = "Yes", title = "Fuel consumption and aspects of car design and performance for various cars",
                            creator = "An organisation")
                            
accessibletablesR::creatingtables(title = "Fuel consumption and aspects of car design and performance for various cars C",
                                  subtitle = "Cars",
                                  extraline1 = "Link to contents",
                                  extraline2 = "Link to notes",
                                  extraline3 = "Link to definitions",
                                  sheetname = "Table_3", table_data = dummydf, tablename = "thirdtable", headrowsize = 40,
                                  numdatacols = c(2:8,11:13), numdatacolsdp = c(1,0,1,0,2,1,2,0,0,3),
                                  othdatacols = c(9,10), columnwidths = "specified",
                                  colwid_spec = c(18,18,18,15,17,15,12,17,12,13,23,22,12))
                                  
accessibletablesR::contentstable()

accessibletablesR::addnote(notenumber = "note1", notetext = "Google is an internet search engine", applictabtext = "All", linktext1 = "Google", linktext2 = "https://www.ons.google.co.uk") 

accessibletablesR::notestab()

accessibletablesR::adddefinition(term = "Usual resident", definition = "A usual resident is anyone who, on Census Day, 21 March 2021 was in the UK and had stayed or intended to stay in the UK for a period of 12 months or more, or had a permanent UK address and was outside the UK and intended to be outside the UK for less than 12 months.")

accessibletablesR::definitionstab()

accessibletablesR::coverpage(title = "Fuel consumption and aspects of car design and performance for various cars",
                             intro = "Some made up data about cars",
                             about = "The output of an example of how to use accessibletablesR",
                             source = "R mtcars",
                             relatedlink = "https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/mtcars)",
                             relatedtext = "mtcars: Motor trend car road tests",
                             dop = "26 October 2023",
                             blank = "There should be no blank cells",
                             names = "Your name",
                             email = "yourname@emailprovider.com",
                             phone = "01111 1111111111111",
                             reuse = "Yes", govdept = NULL)
                             
accessibletablesR::savingtables("D:/mtcarsexample.xlsx", odsfile = "Yes", deletexlsx = "No")

}
